Não deletar planos com detalhes

+++ app/Http/Controllers/Admin/PlanController.php
@@ -55,10 +55,18 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository
+                        ->with('details')
+                        ->where('url', $url)
+                        ->first();
 
         if (!$plan)
             return redirect()->back();
+
+        if ($plan->details->count() > 0)
+            return redirect()
+                        ->back()
+                        ->with('error', 'Exclua os detalhes do plano antes de prosseguir');
 
         $plan->delete();
